requirements.txt
# Core Data Science Libraries
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter and Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Configuration and Utilities
PyYAML>=6.0
joblib>=1.0.0
tqdm>=4.62.0

# Model Explainability
shap>=0.40.0
lime>=0.2.0

# Web Framework and API
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Streamlit for UI
streamlit>=1.0.0

# HTTP Requests
requests>=2.25.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Optional: Advanced ML libraries
# xgboost>=1.5.0
# lightgbm>=3.2.0
# catboost>=1.0.0

# Optional: Deep Learning (uncomment if needed)
# torch>=1.9.0
# tensorflow>=2.6.0

# Optional: Additional utilities
# psutil>=5.8.0  # For system monitoring
# memory-profiler>=0.60.0  # For memory profiling
